+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# WIP
+
+Added:
+
+- macOS: text/plain clipboard support #225 #51 via @bumfo @lgorence
+
 # 0.4.1 - April 6, 2022
 
 Changed: